requirements-nodriver.txt
# ZipRecruiter: undetected Chrome via nodriver (bypass Cloudflare, export cookies for Node)
nodriver>=0.28
python-dotenv>=1.0.0
